pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voicecomment"
version = "0.1.0"
description = "Record voice and insert as inline code comments across any language and IDE"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "VoiceComment Team"}
]
keywords = ["voice", "comments", "code", "transcription", "whisper", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "pyaudio>=0.2.13",
    "numpy>=1.24.0",
    "openai-whisper>=20231117",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
voicecomment = "voicecomment.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/voicecomment"
Repository = "https://github.com/yourusername/voicecomment"
Issues = "https://github.com/yourusername/voicecomment/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["voicecomment*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
